ASM-4621 use NICs in UI-specified order if available
Allow network configuration data to specify the exact NIC to use for
an interface rather than always picking it based on type and number of
partitions.

Additionally add the chosen NIC fqdd to the interface so
that the resulting data is consistent whether the network
configuration specified the NIC or it was chosen by add_nics!