ASM-7091 Use correct attribute for attach mode
The original code for setting attach mode was setting the incorrect
iDrac card enumeration enumeration attribute. This meant that the
server was always left in auto attach mode which requires an
additional reboot in order for the virtual CD to show up in the boot
order list after we attach the ISO. In attach mode the virtual CD is
always present in the boot order list even if no ISO is attached.

In addition to slowing down the install process the additional reboot
was occasionally causing servers to get rebooted while an OS
installation was already in progress and corrupting that install.