ESTS-139864 support ESXi 6.7 deployment
Since ESXi 6.7 requires CA cert installation or thumbprint validation,
lib/asm/util.rb file was updated to meet this requirement.

For the purpose of this task, temporary thumbprint validation seemed
as the most appropriate solution, as this method does not require
clean up during the teardown.

A helper method was implemented to grab and parse the thumbprint,
and it is called from the esxcli method to set a command argument.

Update:
1. Modified esxcli method to get thumbprint only when it is not
   already provided through endpoint[:thumbprint].
2. Removed helper method and moved its logic into esxcli method.
3. Unit test was modified.