ESTS-140275 Allow latest versions of gems in gemspec
Open up most library version requirements in gemspec. This makes it
easier for the consuming application to specify the versions they
require. This is at the expense of making it more likely that
dell-asm-util tests will fail at some point as new library versions
are released.

Rubocop was left at version 0.37.2 because upgrading that will require
updating code style to match and not fail the new rubocop
checks. Puppetlabs_spec_helper was left at 0.4.1 for now as otherwise
rake task fail.

Travis CI was updated to test against jruby 1.7, jruby 9 and ruby
2.4.3 which are the primary targets now