VXFM-2511 [Security] Replace URI.escape with CGI::escape
A recent fortify scan of asm-deployer showed 4 criticals that were all
instances of using URI.escape. That method is marked as "obsolete and
should not be used" in the API docs, although it does not explicitly
say its a security risk: https://docs.ruby-lang.org/en/2.5.0/URI/Escape.html

This fix replaces all occurences of URI.escape with CGI.escape in non-test code.