VXFM-5119 Upgrade to latest nokogiri and rubocop
The versions of nokogiri and rubocop used had low to moderate
security vulnerabilities reported. Nokogiri was previously pinned to a
lower version because of a memory leak with jruby and the 1.8.x
versions. That issue appears to be resolved in the latest
versions. Rubocop was pinned just to avoid failing new checks that
constantly get added. This change updates to the latest versions to
resolve the security warnings.

Most of the changes here are to conform to new Rubocop checks and
update the base ruby version to 2.3 Support for JRuby 1.7 was removed.